Added save iterations to Full_CI_ZMQ (#203)
* Added/Updated files to save iterations 

Modified EZFIO.cfg to include iteratively saved data if the keyword of "full_ci_zmq/iterative" is set to true in the ezfio. The default is false. 
Will save the number of total iterations in full_ci_zmq/n_iter
Saves the number of determinants in full_ci_zmq/n_det_iter
Saves the energy in full_ci_zmq/energy_iter.
Saves the energy_pt2  in full_ci_zmq/energybefore_pt2_iter
These results are the same as the output of the program at every iteration.

Modified fci_zmq.irp.f to include calls to fci_iterations.irp.f at each iteration (starting at N_det==1 and including the final call to do the final pt2 correction)

Created fci_iterations as a subroutine to save the number of determinants, energy, and energy+pt2 for every iteration and saves the results in the full_ci_zmq output directory.

* Updated files to include 3 save options 

Updated "iterative" in EZFIO.cfg to be an integer that can take the values of 1 (Append), 2 (Overwrite), or 3 (Do not save). 
Updated fci_iterations to be simpler and include the three options.

Also updated the output so that only N_det, energy, and pt2 are output. The user must manipulate from there.

* Delete fci_iterations.irp.f

The file was modified and moved to fci_iterations.f90

* Fixed the comments

* Rename fci_iterations to dump_fci_iterations_value

Changed name to be clear on purpose.

* Updated files for iterative saving

Renamed fci_iterations to dump_fci_iterations_value for clarity
In EZFIO.cfg changed keyword "iterative" to "iterative_save" for clarity
Update dump_fci_iterations_value with "iterative_save" keyword
Removed call to dump_fci_iterations on line 27 in fci_zmq
Updated fci_zmq with the name change for calls

* Delete fci_zmq.irp.f

* Delete fci_zmq_pt2.irp.f

* Delete EZFIO.cfg

* Get latest updates from master

* Added calls to dump_fci_iterations_value

* Readded dump_fci_iterations_value

* Updated EZFIO.cfg